requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
playwright==1.37.0
pluggy==1.3.0
pyee==9.0.4
pytest==7.4.1
pytest-base-url==2.0.0
pytest-html==4.0.0
pytest-metadata==3.0.0
pytest-playwright==0.4.2
python-slugify==8.0.1
requests==2.31.0
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.0.4
